the final API for Operation: `Operation.(params, *dependencies)`
i decided that it's important to not mix params with arbitrary incoming dependencies.
`dependencies` will usually be a hash such as `user_repository: User::Repository`.
the cool part here is that dependencies can be a container, such as a `Dry::Container` instance.

the entire *args chain is passed to both Builder and #initialize.
to #call, only `params` is passed as the "runtime" method doesn't need the dependencies.